Eliminate opzioni non richieste

+++ Esame/src/it/unibs/pgar/esame/Entita.java
@@ -6,7 +6,7 @@ public interface Entita {
 	public static int difesaIniziale = 5;
 	
 	public double attacca();
-	public void subisciDanni(int danni);
+	public void subisciDanni(double danni);
 	public boolean inVita();
 
 }
